% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{load_file}
\alias{load_file}
\title{Read .csv/.tsv files or error gracefully}
\usage{
load_file(file)
}
\arguments{
\item{file}{A string giving the path to a file.}
}
\value{
A data frame.
}
\description{
Read .csv/.tsv files or error gracefully
}
\examples{
# Setup
df <- tibble::tibble(x = 1, y = "a")
file <- tempfile(fileext = ".tsv")
vroom::vroom_write(df, file, delim = "\t")

load_file(file)
}
