% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_penguins.R
\name{scatterplot_penguins}
\alias{scatterplot_penguins}
\alias{penguins_labs}
\title{Scatterplot bill dimensions of penguins}
\usage{
scatterplot_penguins(data)

penguins_labs(title = NULL, x = NULL, y = NULL, ...)
}
\arguments{
\item{data}{Data frame structured as \link[palmerpenguins:penguins]{palmerpenguins::penguins}.}

\item{title}{The text for the title.}

\item{x, y}{Labels for the \code{x} and \code{y} axes.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:labs]{ggplot2::labs}}
  \describe{
    \item{\code{subtitle}}{The text for the subtitle for the plot which will be
displayed below the title.}
    \item{\code{caption}}{The text for the caption which will be displayed in the
bottom-right of the plot by default.}
    \item{\code{tag}}{The text for the tag label which will be displayed at the
top-left of the plot by default.}
  }}
}
\description{
Bill length and depth for Adelie, Chinstrap and Gentoo Penguins at Palmer
Station LTER (https://allisonhorst.github.io/palmerpenguins).
}
\examples{
data <- na.omit(palmerpenguins::penguins)

p <- scatterplot_penguins(data)
p

p + penguins_labs(title = "Custom title")
}
